fix: Recommendation Methods
